requirements.txt
cryptography>=41.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
pycryptodome>=3.19.0
colorama>=0.4.6
pyreadline3>=3.4.1
tqdm>=4.45.0 